# docs/openapi.yaml
components:
  schemas:
...
    CoreBuildResponse:
      additionalProperties: false
      properties:
        created_at:
          type: string
        id:
          format: int64
          type: integer
        parent_build_id:
          format: int64
          type: integer
        state:
          type: string
        upload_id:
          format: int64
          type: integer
        user_version:
          type: string
      required:
        - id
        - user_version
        - state
        - created_at
      type: object
...
    WharfBuildFileResponse:
      additionalProperties: false
      properties:
        createdAt:
          type: string
        id:
          format: int64
          type: integer
        size:
          format: int64
          type: integer
        state:
          enum:
            - uploading
            - uploaded
            - failed
            - evicted
            - rebuilding
          type: string
        subType:
          examples:
            - default
          type: string
        type:
          enum:
            - archive
            - patch
            - signature
          type: string
        updatedAt:
          type: string
      required:
        - id
        - size
        - state
        - type
        - subType
        - createdAt
        - updatedAt
      type: object
    WharfBuildResponse:
      additionalProperties: false
      properties:
        createdAt:
          type: string
        files:
          items:
            $ref: "#/components/schemas/WharfBuildFileResponse"
          type:
            - array
            - "null"
        id:
          format: int64
          type: integer
        parentBuild:
          $ref: "#/components/schemas/WharfParentBuildResponse"
        parentBuildId:
          format: int64
          type: integer
        state:
          enum:
            - started
            - processing
            - completed
            - failed
          type: string
        updatedAt:
          type: string
        uploadId:
          format: int64
          type: integer
        userVersion:
          type: string
        version:
          format: int64
          type: integer
      required:
        - id
        - uploadId
        - parentBuildId
        - parentBuild
        - version
        - state
        - userVersion
        - files
        - createdAt
        - updatedAt
      type: object
...
    WharfParentBuildResponse:
      additionalProperties: false
      properties:
        id:
          format: int64
          type: integer
      required:
        - id
      type: object
...
  securitySchemes:
    ApiKeyAuth:
      description: Raw API key, Bearer token, or access_token token.
      in: header
      name: Authorization
      type: apiKey
info:
  title: DenKit Stash API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /builds/{id}:
    get:
      operationId: get-build
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  build:
                    $ref: "#/components/schemas/CoreBuildResponse"
                required:
                  - build
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: Get a build by ID
      tags:
        - Core
...
  /uploads/{id}/builds:
    get:
      operationId: list-upload-builds
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: int64
            type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  builds:
                    items:
                      $ref: "#/components/schemas/CoreBuildResponse"
                    type:
                      - array
                      - "null"
                required:
                  - builds
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
        - {}
      summary: List builds for an upload
      tags:
        - Core
  /wharf/builds:
    get:
      operationId: list-wharf-builds
      parameters:
        - description: Target in username/game form
          in: query
          name: target
          required: true
          schema:
            type: string
        - description: Channel name
          in: query
          name: channel
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  builds:
                    items:
                      $ref: "#/components/schemas/CoreBuildResponse"
                    type:
                      - array
                      - "null"
                required:
                  - builds
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
      security:
        - ApiKeyAuth: []
      summary: List builds for a target channel
      tags:
        - Wharf
    post:
      operationId: create-wharf-build
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: false
              properties:
                channel:
                  examples:
                    - main
                  type: string
                target:
                  examples:
                    - username/game
                  type: string
                user_version:
                  type: string
              required:
                - target
                - channel
              type: object
          application/x-www-form-urlencoded:
            schema:
              additionalProperties: false
              properties:
                channel:
                  examples:
                    - main
                  type: string
                target:
                  examples:
                    - username/game
                  type: string
                user_version:
                  type: string
              required:
                - target
                - channel
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  build:
                    $ref: "#/components/schemas/WharfBuildResponse"
                required:
                  - build
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Not Found
        "413":
          content:
            application/json:
              schema:
                additionalProperties: false
                properties:
                  errors:
                    items:
                      type: string
                    type:
                      - array
                      - "null"
                required:
                  - errors
                type: object
          description: Request Entity Too Large
      security:
        - ApiKeyAuth: []
      summary: Create a build for a target channel
      tags:
        - Wharf
...